Update benchmarks & readme
